Moved client-sided code into systems folder.
