+++ systems/generic/client/character.scss
